Fix types for stat struct
When running the same program using the stat
structure in Linux and with Unikraft the sizes
of the structure's fields were different. This was
caused by the fact that in Linux dev_t was declared
as unsigned long, while in Unikraft it was declared
as __u64. Similarly for ino_t, nlink_t, blkcnt_t.

+++ lib/nolibc/include/nolibc-internal/shareddefs.h
@@ -99,22 +99,22 @@ typedef unsigned id_t;
 #endif
 
 #if defined(__NEED_dev_t) && !defined(__DEFINED_dev_t)
-typedef __u64 dev_t;
+typedef unsigned long dev_t;
 #define __DEFINED_dev_t
 #endif
 
 #if defined(__NEED_ino_t) && !defined(__DEFINED_ino_t)
-typedef __u64 ino_t;
+typedef unsigned long ino_t;
 #define __DEFINED_ino_t
 #endif
 
 #if defined(__NEED_nlink_t) && !defined(__DEFINED_nlink_t)
-typedef __u32 nlink_t;
+typedef unsigned long nlink_t;
 #define __DEFINED_nlink_t
 #endif
 
 #if defined(__NEED_blkcnt_t) && !defined(__DEFINED_blkcnt_t)
-typedef __s64 blkcnt_t;
+typedef long blkcnt_t;
 #define __DEFINED_blkcnt_t
 #endif
